requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
alembic
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
python-multipart
python-dateutil
pandas
openpyxl
numpy
datetime
xlsxwriter
reportlab
cryptography